Add files via upload
The primary objective of this project was to create a holistic solution. It began with the ingestion of tables from an on-premise SQL Server database, orchestrated by Azure Data Factory. The data was then seamlessly stored within Azure Data Lake. Subsequently, Azure Databricks came into play, enabling the transformation of raw data into its most pristine and valuable form. The final piece of the puzzle was Microsoft Power BI, which integrated with Azure Synapse Analytics, resulting in the development of an interactive dashboard.

Hosted Link listed

https://app.powerbi.com/view?r=eyJrIjoiMjc5MDU3YWYtNGUzMC00MWQ2LTlmNWEtYjNiNDUzM2U0ZTU2IiwidCI6IjQwMzljZmY3LTAwNTItNDU2NC04NmVlLWRkY2YzNGZhN2U0ZSIsImMiOjZ9